requirement.txt
Package      Version  
------------ ---------
baidu-aip    2.2.17.0     
PyAudio      0.2.11   
pynput       1.4.2    
Wave         0.0.2    
